+++ backend/src/webhooks/webhooks.module.ts
@@ -1,9 +1,0 @@
-import { Module } from "@nestjs/common";
-import { WebhooksController } from "./webhooks.controller";
-import { KycModule } from "../kyc/kyc.module";
-
-@Module({
-  imports: [KycModule],
-  controllers: [WebhooksController],
-})
-export class WebhooksModule {}